# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HouseResponse": {
        "description": "DTO for returning house details.",
        "properties": {
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "area": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "bathrooms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bathrooms"
          },
          "city": {
            "title": "City",
            "type": "string"
          },
          "community": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Community"
          },
          "condo_fee": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condo Fee"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "days_on_market": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days On Market"
          },
          "decoration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decoration"
          },
          "floor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "images": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Images"
          },
          "is_synthetic": {
            "default": false,
            "title": "Is Synthetic",
            "type": "boolean"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "listed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listed At"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "original_price": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Price"
          },
          "parking": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parking"
          },
          "postal_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postal Code"
          },
          "price": {
            "title": "Price",
            "type": "integer"
          },
          "price_per_sqft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per Sqft"
          },
          "property_tax": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Tax"
          },
          "property_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Type"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "rooms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rooms"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "sqft": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sqft"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "active",
            "title": "Status"
          },
          "street": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Street"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "title",
          "city",
          "region",
          "price",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "HouseResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Canadian residential investment analytics: listings, comparables, cash flow, markets.",
    "title": "NeighborIQ API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/communities": {
      "get": {
        "description": "List neighbourhoods with aggregated statistics.",
        "operationId": "list_communities_api_v1_communities_get",
        "parameters": [
          {
            "description": "Filter by city",
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by city",
              "title": "City"
            }
          },
          {
            "description": "Filter by region",
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by region",
              "title": "Region"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Communities Api V1 Communities Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Communities",
        "tags": [
          "listings"
        ]
      }
    },
    "/api/v1/communities/{community_id}/stats": {
      "get": {
        "description": "Get aggregated statistics for a neighbourhood.",
        "operationId": "get_community_stats_api_v1_communities__community_id__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "community_id",
            "required": true,
            "schema": {
              "title": "Community Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Community Stats Api V1 Communities  Community Id  Stats Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Community Stats",
        "tags": [
          "listings"
        ]
      }
    },
...
    "/api/v1/houses/search": {
      "get": {
        "description": "Quick search (typeahead): up to 50 matches by keyword and location.",
        "operationId": "search_houses_api_v1_houses_search_get",
        "parameters": [
          {
            "description": "Full-text search query",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Full-text search query",
              "title": "Q"
            }
          },
          {
            "description": "Filter by city",
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by city",
              "title": "City"
            }
          },
          {
            "description": "Filter by region",
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by region",
              "title": "Region"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Houses",
        "tags": [
          "listings"
        ]
      }
    },
    "/api/v1/houses/{house_id}": {
      "delete": {
        "description": "Soft-delete a listing (admin only).",
        "operationId": "delete_house_api_v1_houses__house_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "house_id",
            "required": true,
            "schema": {
              "title": "House Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete House Api V1 Houses  House Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete House",
        "tags": [
          "listings"
        ]
      },
      "get": {
        "description": "Get listing details by ID.",
        "operationId": "get_house_api_v1_houses__house_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "house_id",
            "required": true,
            "schema": {
              "title": "House Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get House",
        "tags": [
          "listings"
        ]
      },
      "put": {
        "description": "Partially update a listing (admin only). Price changes are recorded in price history.",
        "operationId": "update_house_api_v1_houses__house_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "house_id",
            "required": true,
            "schema": {
              "title": "House Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HouseUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update House",
        "tags": [
          "listings"
        ]
      }
    },
...
    "/api/v1/houses/{house_id}/neighbourhood": {
      "get": {
        "description": "Nearest schools, transit stops and hospitals (OpenStreetMap, linked by the\ningestion `osm` loader). Empty lists mean POIs have not been loaded yet.",
        "operationId": "get_neighbourhood_api_v1_houses__house_id__neighbourhood_get",
        "parameters": [
          {
            "in": "path",
            "name": "house_id",
            "required": true,
            "schema": {
              "title": "House Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Neighbourhood Api V1 Houses  House Id  Neighbourhood Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Neighbourhood",
        "tags": [
          "listings"
        ]
      }
    },
    "/api/v1/houses/{house_id}/price-history": {
      "get": {
        "description": "Asking-price changes over time, oldest first.",
        "operationId": "get_price_history_api_v1_houses__house_id__price_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "house_id",
            "required": true,
            "schema": {
              "title": "House Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Price History Api V1 Houses  House Id  Price History Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Price History",
        "tags": [
          "listings"
        ]
      }
    },
...
  }
}